Wheat yields

In this field trial, the yields of a standard commercial wheat variety and a new ‘improved’ variety were compared to assess differences in average performance. Seven small plots were planted with each variety, resulting in yield measurements from a total of 14 plots. Yields were determined for each plot and standardized to tonnes per hectare (t/ha). The primary objective of the study was to evaluate whether the new variety offers a significant advantage in yield over the standard commercial type.

Usage

wheat

Format

A data frame with 3 variables: DPlot, Variety, Yield.

DPlot

Factor. Unique identifier for each plot in the field trial.

Variety

Factor. Wheat variety grown in each plot ("Commercial" or "Improved").

Yield

Numeric. Grain yield from the plot, measured in tonnes per hectare (t/ha).

Source

Welham, S. J., Gezan, S. A., Clark, S. J., and Mead, A. (2015) Statistical Methods in Biology: Design and analysis of experiments and regression

Examples

lm(Yield ~ Variety, data = wheat)
#> 
#> Call:
#> lm(formula = Yield ~ Variety, data = wheat)
#> 
#> Coefficients:
#>     (Intercept)  VarietyImproved  
#>           7.571            2.714  
#>